requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.0
numpy==1.19.4
Pillow==7.0.0
requests==2.22.0
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp38-cp38-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.8.2%2Bcpu-cp38-cp38-linux_x86_64.whl
urllib3==1.25.8
Werkzeug==1.0.1
